# Optimize lub algorithm