@odersky 2024-03-27
Based on #20027. Last commit is new.