@noti0na1 2024-03-27
test performance please